Remove Gemini integration and add simple Slack bot
